backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
groq
python-dotenv
python-multipart
websockets
pydantic
